[ion/timing] millis return uint64_t
Otherwise we could loop after a month or so

+++ ion/src/device/boot/rt0.cpp
@@ -93,7 +93,7 @@ void start() {
   abort();
 }
 
-volatile uint32_t millis_elapsed = 0;
+volatile uint64_t millis_elapsed = 0;
 
 void __attribute__((interrupt)) sysTick() {
   millis_elapsed++;
